pyproject.toml
[tool.poetry]
name = "contact-manager"
version = "0.1.0"
description = "Contact Management System that aggregates contacts from various sources"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<3.14"
pyside6 = { version = "^6.5.0", optional = true }
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
aiohttp = "^3.8.0"
google-auth-oauthlib = "^1.0.0"
rapidfuzz = "^3.6.1"
pydantic = "^2.0.0"
aiosqlite = "^0.19.0"
google-api-python-client = "^2.0.0"
python-dotenv = "^1.0.1"
qasync = "^0.27.0"
authlib = "^1.3.2"
vobject = "^0.9.6.1"
caldav = "^1.3.6"
requests = "^2.32.3"
greenlet = "^3.0.0"
fastapi = "^0.104.0"
uvicorn = "^0.24.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.6"
google-auth-httplib2 = "^0.2.0"
psutil = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0"
black = "^23.0.0"

[tool.poetry.extras]
desktop = ["pyside6"]
web = ["fastapi", "jinja2"]